src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-on-image-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: imageDiv.createShadowRoot is not a function
